# swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "Customer API",
      "description": "API for managing customers"
    },
    "host": "localhost:4000",
    "basePath": "/",
    "schemes": ["http"],
    "paths": {
      "/addCustomer": {
        "post": {
          "summary": "Add a new customer",
          "description": "Creates a new customer in the system",
          "parameters": [
            {
              "in": "body",
              "name": "customer",
              "description": "Customer to add",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "example": "John Doe" },
                  "email": { "type": "string", "example": "johndoe@example.com" },
                  "password": { "type": "string", "example": "secret" }
                },
                "required": ["name"]
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Customer created successfully",
              "schema": {
                "$ref": "#/definitions/Customer"
              }
            },
            "400": {
              "description": "Bad request"
            }
          }
        }
      },
      "/getAll": {
        "get": {
          "summary": "Get all customers",
          "description": "Returns a list of all customers",
          "responses": {
            "200": {
              "description": "A list of customers",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Customer"
                }
              }
            },
            "500": {
              "description": "Server error"
            }
          }
        }
      },
      "/getCustomer/{id}": {
        "get": {
          "summary": "Get customer by ID",
          "description": "Returns a customer by their ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "type": "string",
              "description": "UUID of the customer"
            }
          ],
          "responses": {
            "200": {
              "description": "Customer found",
              "schema": {
                "$ref": "#/definitions/Customer"
              }
            },
            "404": {
              "description": "Customer not found"
            },
            "500": {
              "description": "Server error"
            }
          }
        }
      },
      "/updateCustomer/{id}": {
        "put": {
          "summary": "Update customer by ID",
          "description": "Updates a customer's details by their ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "type": "string",
              "description": "UUID of the customer"
            },
            {
              "in": "body",
              "name": "customer",
              "description": "Customer data to update",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Customer updated successfully",
              "schema": {
                "$ref": "#/definitions/Customer"
              }
            },
            "400": {
              "description": "Invalid UUID format"
            },
            "404": {
              "description": "Customer not found"
            },
            "500": {
              "description": "Server error"
            }
          }
        }
      },
      "/deleteCustomer/{id}": {
        "delete": {
          "summary": "Delete customer by ID",
          "description": "Deletes a customer by their ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "type": "string",
              "description": "UUID of the customer"
            }
          ],
          "responses": {
            "200": {
              "description": "Customer deleted successfully"
            },
            "404": {
              "description": "Customer not found"
            },
            "500": {
              "description": "Server error"
            }
          }
        }
      }
    },
    "definitions": {
      "Customer": {
        "type": "object",
        "properties": {
          "customerId": { "type": "string" },
          "name": { "type": "string" },
          "email": { "type": "string" },
          "password": { "type": "string" }
        }
      }
    }
  }